backend/requirements.txt
alembic==1.13.1
anyio==4.3.0
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
click==8.1.7
Deprecated==1.2.14
distlib==0.3.8
filelock==3.13.1
Flask==3.0.2
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
platformdirs==4.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
svix==1.21.0
types-Deprecated==1.2.9.20240311
types-python-dateutil==2.8.19.20240311
typing_extensions==4.10.0
virtualenv==20.25.1
Werkzeug==3.0.1
wrapt==1.16.0
